remove const modifier on Tile::index
As a relative C++ newbie, the behaviour for implicit function deletion
really killed me for the past hour or two. The amount of stuff that
goes on behind the scenes (and can be pulled from under your feet from
behind the scenes) is kind of mind boggling. I'm not deep enough into
the language to deal with this right now.