fix array size bug when printing backtrack heatmap
Today I learned that c compilers are not supposed to warn if an array
initializer list is not the size specified in the template argument.
That's a killer nuance. I probably should have noticed on my own, but
I'm used to static checkers and compilers telling me about these kinds
of things :P