change newline placement convention to: leading
Rationale: It looks nicer when printing lines of a help menu, or of
various statistics (They become part of string literals on the left
of the line of source code as opposed to a sudden char literal on
the right of lines of source code that are not the same length). It
also means I have to think less. Basically, whenever there is some-
thing that goes on its own line, it's that thing's job to emit a new
line to go before itself. It doesn't need to worry about whether the
code before it prepared one for it (it may not have! consider how I
print trials.), and when it comes to creating space between sections
of less-related text, then each section prepares that extra space
for whatever comes next. The opposite convention of preparing new-
lines at the end of each line is a habit I don't like that has stuck
from the first class I took to be introduced to programming.