remove puzzle-solving capabilities
Looking back, I consider this as feature creep. My original goal did
not place emphasis on solving puzzles. It just happened that it was
something that was posible to do. The algorithm's ability to solve
puzzles was nothing impressive anyway. Taking these things out removes
a lot of complexity. I'm happy to do this.