remove the Tile.value field for perf gain
The value is only used for final display and erasing has_masks. It
can be recovered by using the try_order array and index of the Tile.
The reduction in space and removal of instructions needed to
maintain the field resulted in a performance gain:

An a/b test with the same random seed and the command `gen_ok 30000`
showed time elapsed dropping by 1/5 of the original value.