add examples and try to get complete template argument deduction
Ran into problems related to variant-of-span mapping back to layout of
variant-of-array and how to change the order of the backed variant-of-array
only given the variant-of-span.

Thought hard about why I made GridSpan and GridArr in the first place
and tried to come to a comclusion on whether that separation is actually
valuable. You can find that in the TODO.md file. The conclusion: I
currently cannot think of a good reason for that separation. I think I
should get rid of it and make my life easier.

I learned some CTAD stuff and attempted to read the standardese on
template deduction for functions and classes. Bit of a shame that it
seems to now be going to waste. Maybe it will be useful later though.