flatten some namespaces
Rationale: most of the the library is just working with a few exposed classes.
Some of the namespaces I had were not very easy to argue for the value.
The archive stuff I think makes sense to have a namespace since it will have
operations as free functions on its own specific classes. I haven't flattened
the folders corresponding to the now-flattened namespaces. I know this is a
violation of the pitchfork layout conventions. I'm not sure what to do. I feel
like the folder layout is still useful in grouping things conceptually.