improve random grid generation performance
changed to use mod arithmetic to avoid PRNG of same number instead of
`continue`ing the loop when second number is the same.

also added `final` back to most classes/structs. was previously trying
to follow C++ core guidelines, but prefer now something narrow that I can
open up in the future if/when needed.

and moved OKIIDOKU_DEFER_INIT to go after variable name instead of before
type. GCC and Clang don't seem to care which of those positions the
attribute goes, but cppreference shows `[[indeterminate]]` after var name
so I'll just put it there. and it makes it easier to read I think.