misc improvements + found grid MRUIS bug
copying TODO comment:
for the last read_buf, reader needs to know how many bytes need to be read and not
overread. it worked when I tested only one logical MRUI to a stream because reading
would hit EOF. there are various ways I could address this (ex. reserving a bit per
byte to signal whether the byte is the last one, or making a zero- byte instead be
a zero byte plus a following byte with such a signal bit, or by making the caller
give the reader the length of data to read up-front).

I had been assuming that modulo math worked in a way that... it just doesn't. sigh.